requirements.txt
aiofiles==0.4.0
asyncio==3.4.3
certifi==2019.9.11
chardet==3.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
multidict==4.5.2
pyaes==1.6.1
pyasn1==0.4.7
python-dotenv==0.10.3
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
rsa==4.0
sanic==19.6.3
Telethon==1.10.3
urllib3==1.25.5
websockets==7.0
